GEO-INFER-GIT/requirements.txt
# GEO-INFER-GIT Requirements
# Core dependencies
requests>=2.28.1
pyyaml>=6.0
psutil>=5.9.0

# Development dependencies
pytest>=7.3.1
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
tqdm>=4.65.0
colorlog>=6.7.0
pytest-cov>=4.1.0
isort>=5.12.0 